- vs .sln file created. - created Utils and Tests dedicated folders - files moved.
