+++ Data-Structures.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.4.33122.133
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Data-Structures", "Data-Structures.vcxproj", "{CF3AD935-6924-456B-8ECB-3A6C85FD6256}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CF3AD935-6924-456B-8ECB-3A6C85FD6256}.Debug|x64.ActiveCfg = Debug|x64
+		{CF3AD935-6924-456B-8ECB-3A6C85FD6256}.Debug|x64.Build.0 = Debug|x64
+		{CF3AD935-6924-456B-8ECB-3A6C85FD6256}.Debug|x86.ActiveCfg = Debug|Win32
+		{CF3AD935-6924-456B-8ECB-3A6C85FD6256}.Debug|x86.Build.0 = Debug|Win32
+		{CF3AD935-6924-456B-8ECB-3A6C85FD6256}.Release|x64.ActiveCfg = Release|x64
+		{CF3AD935-6924-456B-8ECB-3A6C85FD6256}.Release|x64.Build.0 = Release|x64
+		{CF3AD935-6924-456B-8ECB-3A6C85FD6256}.Release|x86.ActiveCfg = Release|Win32
+		{CF3AD935-6924-456B-8ECB-3A6C85FD6256}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {17B69DB9-C30A-43A9-B995-700C4F53235F}
+	EndGlobalSection
+EndGlobal
