+++ main.cpp
@@ -1,8 +1,8 @@
-#include "DetectPlatform.h"
+#include "Utils/DetectPlatform.h"
 
-#include "testList.hpp"
-#include "testSimpleList.hpp"
-#include "testBinaryTree.hpp"
+#include "Tests/testList.hpp"
+#include "Tests/testSimpleList.hpp"
+#include "Tests/testBinaryTree.hpp"
 
 int main()
 {
